// src/types/enum.ts
export enum NavLinkEnum {
  introduction = "/introduction",
  skills = "/skills",
  socials = "/socials",
  support = "/support",
}

export enum ServiceLinkEnum {
  github = "https://github.com/",
  linkedin = "https://www.linkedin.com/in/",
  twitter = "https://twitter.com/",
  instagram = "https://www.instagram.com/",
  facebook = "https://www.facebook.com/",
  youtube = "https://www.youtube.com/",
  buyMeACoffee = "https://www.buymeacoffee.com/",
  kofi = "https://ko-fi.com/",
}